Crosswalking between linkml and bioschemas
This day-2 task investigates using "exact_mapping" attributes in LinkML
models as the basis for cross walking to bioschemas. Currently, these
exact mappings are encoded in the source yaml but future versions might
choose to use an extra crosswalking file like
https://github.com/linkml/rda-crosswalk/blob/main/crosswalk.yaml to
achieve the same thing in a more independent fashion.

The current code only does the parsing of the necessary information.
Choices need to be made on the formatting of the BSO JSON-LD.